refactor: RacingCar -> CarRacing 네이밍 수정
