+++ src/main/java/racingcar/constant/CarRacingErrorMessage.java
@@ -1,6 +1,6 @@
 package racingcar.constant;
 
-public class RacingCarErrorMessage {
+public class CarRacingErrorMessage {
 
     public static final String CAR_NAME_LENGTH_ERROR_MESSAGE = "자동차 이름은 5자를 초과할 수 없습니다.";
     public static final String TRY_COUNT_NUMBER_FORMAT_ERROR_MESSAGE = "이동 횟수는 숫자값만 입력 가능합니다.";
