+++ src/main/java/racingcar/constant/CarRacingMessage.java
@@ -1,6 +1,6 @@
 package racingcar.constant;
 
-public class RacingCarMessage {
+public class CarRacingMessage {
 
     public static final String RACING_RESULT_MESSAGE = "\n실행 결과";
     public static final String INPUT_CAR_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
