+++ src/main/java/racingcar/domain/CarRacing.java
@@ -1,16 +1,17 @@
 package racingcar.domain;
 
-import static racingcar.constant.RacingCarMessage.RACING_RESULT_MESSAGE;
+import static racingcar.constant.CarRacingMessage.RACING_RESULT_MESSAGE;
 
 import java.util.List;
+import racingcar.io.CarRacingOutputManager;
 
 public class CarRacing {
     private final Cars cars;
     private final Turn turn;
     private final Referee referee;
-    private final RacingCarOutputManager outputManager;
+    private final CarRacingOutputManager outputManager;
 
-    public CarRacing(Cars cars, Turn turn, Referee referee, RacingCarOutputManager outputManager) {
+    public CarRacing(Cars cars, Turn turn, Referee referee, CarRacingOutputManager outputManager) {
         this.cars = cars;
         this.turn = turn;
         this.referee = referee;
